fix remaining merge issues
